refactor: import TooltipCoreProps from the new types file
